Fix 'show' command  + Changed json file format
